endret current floor til float

+++ source/floor_sensor_logic.c
@@ -8,11 +8,11 @@
 #include "floor_sensor_logic.h"
 
 
-int current_floor = 0;
+float current_floor = 0;
 
-int * current_floor_pointer = &current_floor;
+float * current_floor_pointer = &current_floor;
 
-int new_floor_registered(int * current_floor_pointer){
+int new_floor_registered(float * current_floor_pointer){
 if(hardware_read_floor_sensor(0) && (*current_floor_pointer!=0)){
     *current_floor_pointer = 0;
     hardware_command_floor_indicator_on(0);
